+++ src/useq/_iter_sequence.py
